refreshTotalPrice moved to lib.js
