+++ site/cart/cart-script.js
@@ -1,5 +1,7 @@
 import listener from '../assets/common-scripts/listeners.js';
-import { refreshLoginDependant, refreshAllCardButtons, fetchObject } from '../assets/common-scripts/lib.js';
+import {
+  refreshLoginDependant, refreshAllCardButtons, fetchObject, refreshTotalPrice,
+} from '../assets/common-scripts/lib.js';
 import productList from '../assets/common-scripts/productList.js';
 
 const logoutButton = document.querySelector('.logout-button');
@@ -26,12 +28,5 @@ fetchObject('/api/user/')
     document.querySelector('.username').textContent = resp.loginName;
   });
 // вынести
-const refreshTotalPrice = () => {
-  const totalCostElem = document.querySelector('.total-price');
-  fetch('/api/cart-total-price/')
-    .then((resp) => resp.json())
-    .then(JSON.stringify)
-    .then((totalCost) => { totalCostElem.textContent = `${totalCost} \u20bd`; });
-};
 
 refreshTotalPrice();
